Split out JsonPathField tests into its own test module.
