+++ tests/unit/conditions/test_json_api_condition.py
@@ -2,32 +2,13 @@ import json
 
 import pytest
 import requests
-from marshmallow import ValidationError
 
 from nucypher.policy.conditions.exceptions import (
     InvalidCondition,
     JsonRequestException,
 )
 from nucypher.policy.conditions.json.api import JsonApiCondition
-from nucypher.policy.conditions.json.base import JSONPathField
 from nucypher.policy.conditions.lingo import ConditionLingo, ReturnValueTest
-
-
-def test_jsonpath_field_valid():
-    field = JSONPathField()
-    valid_jsonpath = "$.store.book[0].price"
-    result = field.deserialize(valid_jsonpath)
-    assert result == valid_jsonpath
-
-
-def test_jsonpath_field_invalid():
-    field = JSONPathField()
-    invalid_jsonpath = "invalid jsonpath"
-    with pytest.raises(
-        ValidationError,
-        match=f"'{invalid_jsonpath}' is not a valid JSONPath expression",
-    ):
-        field.deserialize(invalid_jsonpath)
 
 
 def test_json_api_condition_initialization():
